refactor: relocate name-formatting utils into format_utils
Consolidate the mangled-name helpers into `utils/format_utils.py` so they live
next to the other key/name utilities and can be imported without pulling in
`__main__` or `trampoline_templates`:

- Move `CLASS_NAME_SEPARATOR` + `mangle_function_name` out of
  `mutation/trampoline_templates.py` (which no longer referenced them beyond the
  definition) into `format_utils`; `file_mutation` now imports
  `mangle_function_name` from there.
- Move `mangled_name_from_mutant_name` + `orig_function_and_class_names_from_key`
  out of `__main__` into `format_utils`, removing the temporary
  `mangled_name_from_mutant_name` duplicate introduced with the browser helpers.
  `__main__` re-imports both (so `from mutmut.__main__ import ...` still works),
  and `mutation/trampoline.py` now imports the helper from `format_utils`,
  dropping one `__main__` import cycle.

`format_utils` is now a leaf module (stdlib only). Pure move, no behavior change;
full suite is green.
